made more changes to address issues with eslint

+++ src/user/auth.js
@@ -59,32 +59,18 @@ module.exports = function (User) {
         await db.deleteAll([`loginAttempts:${uid}`, `lockout:${uid}`]);
     };
 
-    const getSessionFromStore = util.promisify((sid, callback) =>
-        db.sessionStore.get(sid, (err, sessObj) =>
-            callback(err, sessObj || null),
-        ),
-    );
-    const sessionStoreDestroy = util.promisify((sid, callback) =>
-        db.sessionStore.destroy(sid, err => callback(err)),
-    );
+    const getSessionFromStore = util.promisify((sid, callback) => db.sessionStore.get(sid, (err, sessObj) => callback(err, sessObj || null)));
+    const sessionStoreDestroy = util.promisify((sid, callback) => db.sessionStore.destroy(sid, err => callback(err)));
 
     User.auth.getSessions = async function (uid, curSessionId) {
         await cleanExpiredSessions(uid);
-        const sids = await db.getSortedSetRevRange(
-            `uid:${uid}:sessions`,
-            0,
-            19,
-        );
-        let sessions = await Promise.all(
-            sids.map(sid => getSessionFromStore(sid)),
-        );
+        const sids = await db.getSortedSetRevRange(`uid:${uid}:sessions`, 0, 19);
+        let sessions = await Promise.all(sids.map(sid => getSessionFromStore(sid)));
         sessions = sessions
             .map((sessObj, idx) => {
                 if (sessObj && sessObj.meta) {
                     sessObj.meta.current = curSessionId === sids[idx];
-                    sessObj.meta.datetimeISO = new Date(
-                        sessObj.meta.datetime,
-                    ).toISOString();
+                    sessObj.meta.datetimeISO = new Date(sessObj.meta.datetime).toISOString();
                     sessObj.meta.ip = validator.escape(String(sessObj.meta.ip));
                 }
                 return sessObj && sessObj.meta;
@@ -94,9 +80,7 @@ module.exports = function (User) {
     };
 
     async function cleanExpiredSessions(uid) {
-        const uuidMapping = await db.getObject(
-            `uid:${uid}:sessionUUID:sessionId`,
-        );
+        const uuidMapping = await db.getObject(`uid:${uid}:sessionUUID:sessionId`);
         if (!uuidMapping) {
             return;
         }
@@ -106,22 +90,14 @@ module.exports = function (User) {
             Object.keys(uuidMapping).map(async uuid => {
                 const sid = uuidMapping[uuid];
                 const sessionObj = await getSessionFromStore(sid);
-                const expired =
-                    !sessionObj ||
-                    !sessionObj.hasOwnProperty('passport') ||
-                    !sessionObj.passport.hasOwnProperty('user') ||
-                    parseInt(sessionObj.passport.user, 10) !==
-                        parseInt(uid, 10);
+                const expired = !sessionObj || !sessionObj.hasOwnProperty('passport') || !sessionObj.passport.hasOwnProperty('user') || parseInt(sessionObj.passport.user, 10) !== parseInt(uid, 10);
                 if (expired) {
                     expiredUUIDs.push(uuid);
                     expiredSids.push(sid);
                 }
-            }),
+            })
         );
-        await db.deleteObjectFields(
-            `uid:${uid}:sessionUUID:sessionId`,
-            expiredUUIDs,
-        );
+        await db.deleteObjectFields(`uid:${uid}:sessionUUID:sessionId`, expiredUUIDs);
         await db.sortedSetRemove(`uid:${uid}:sessions`, expiredSids);
     }
 
@@ -135,46 +111,25 @@ module.exports = function (User) {
     };
 
     async function revokeSessionsAboveThreshold(uid, maxUserSessions) {
-        const activeSessions = await db.getSortedSetRange(
-            `uid:${uid}:sessions`,
-            0,
-            -1,
-        );
+        const activeSessions = await db.getSortedSetRange(`uid:${uid}:sessions`, 0, -1);
         if (activeSessions.length > maxUserSessions) {
-            const sessionsToRevoke = activeSessions.slice(
-                0,
-                activeSessions.length - maxUserSessions,
-            );
-            await Promise.all(
-                sessionsToRevoke.map(sessionId =>
-                    User.auth.revokeSession(sessionId, uid),
-                ),
-            );
+            const sessionsToRevoke = activeSessions.slice(0, activeSessions.length - maxUserSessions);
+            await Promise.all(sessionsToRevoke.map(sessionId => User.auth.revokeSession(sessionId, uid)));
         }
     }
 
     User.auth.revokeSession = async function (sessionId, uid) {
-        winston.verbose(
-            `[user.auth] Revoking session ${sessionId} for user ${uid}`,
-        );
+        winston.verbose(`[user.auth] Revoking session ${sessionId} for user ${uid}`);
         const sessionObj = await getSessionFromStore(sessionId);
         if (sessionObj && sessionObj.meta && sessionObj.meta.uuid) {
-            await db.deleteObjectField(
-                `uid:${uid}:sessionUUID:sessionId`,
-                sessionObj.meta.uuid,
-            );
+            await db.deleteObjectField(`uid:${uid}:sessionUUID:sessionId`, sessionObj.meta.uuid);
         }
-        await Promise.all([
-            db.sortedSetRemove(`uid:${uid}:sessions`, sessionId),
-            sessionStoreDestroy(sessionId),
-        ]);
+        await Promise.all([db.sortedSetRemove(`uid:${uid}:sessions`, sessionId), sessionStoreDestroy(sessionId)]);
     };
 
     User.auth.revokeAllSessions = async function (uids, except) {
         uids = Array.isArray(uids) ? uids : [uids];
-        const sids = await db.getSortedSetsMembers(
-            uids.map(uid => `uid:${uid}:sessions`),
-        );
+        const sids = await db.getSortedSetsMembers(uids.map(uid => `uid:${uid}:sessions`));
         const promises = [];
         uids.forEach((uid, index) => {
             const ids = sids[index].filter(id => id !== except);
@@ -190,19 +145,12 @@ module.exports = function (User) {
             'users:joindate',
             async uids => {
                 const sessionKeys = uids.map(uid => `uid:${uid}:sessions`);
-                const sessionUUIDKeys = uids.map(
-                    uid => `uid:${uid}:sessionUUID:sessionId`,
-                );
-                const sids = _.flatten(
-                    await db.getSortedSetRange(sessionKeys, 0, -1),
-                );
+                const sessionUUIDKeys = uids.map(uid => `uid:${uid}:sessionUUID:sessionId`);
+                const sids = _.flatten(await db.getSortedSetRange(sessionKeys, 0, -1));
 
-                await Promise.all([
-                    db.deleteAll(sessionKeys.concat(sessionUUIDKeys)),
-                    ...sids.map(sid => sessionStoreDestroy(sid)),
-                ]);
+                await Promise.all([db.deleteAll(sessionKeys.concat(sessionUUIDKeys)), ...sids.map(sid => sessionStoreDestroy(sid))]);
             },
-            { batch: 1000 },
+            { batch: 1000 }
         );
     };
 };
