+++ src/user/uploads.js
@@ -11,28 +11,18 @@ const file = require('../file');
 const batch = require('../batch');
 
 const md5 = filename => crypto.createHash('md5').update(filename).digest('hex');
-const _getFullPath = relativePath =>
-    path.resolve(nconf.get('upload_path'), relativePath);
+const _getFullPath = relativePath => path.resolve(nconf.get('upload_path'), relativePath);
 const _validatePath = async relativePaths => {
     if (typeof relativePaths === 'string') {
         relativePaths = [relativePaths];
     } else if (!Array.isArray(relativePaths)) {
-        throw new Error(
-            `[[error:wrong-parameter-type, relativePaths, ${typeof relativePaths}, array]]`,
-        );
+        throw new Error(`[[error:wrong-parameter-type, relativePaths, ${typeof relativePaths}, array]]`);
     }
 
     const fullPaths = relativePaths.map(path => _getFullPath(path));
-    const exists = await Promise.all(
-        fullPaths.map(async fullPath => file.exists(fullPath)),
-    );
+    const exists = await Promise.all(fullPaths.map(async fullPath => file.exists(fullPath)));
 
-    if (
-        !fullPaths.every(fullPath =>
-            fullPath.startsWith(nconf.get('upload_path')),
-        ) ||
-        !exists.every(Boolean)
-    ) {
+    if (!fullPaths.every(fullPath => fullPath.startsWith(nconf.get('upload_path'))) || !exists.every(Boolean)) {
         throw new Error('[[error:invalid-path]]');
     }
 };
@@ -40,27 +30,19 @@ const _validatePath = async relativePaths => {
 module.exports = function (User) {
     User.associateUpload = async (uid, relativePath) => {
         await _validatePath(relativePath);
-        await Promise.all([
-            db.sortedSetAdd(`uid:${uid}:uploads`, Date.now(), relativePath),
-            db.setObjectField(`upload:${md5(relativePath)}`, 'uid', uid),
-        ]);
+        await Promise.all([db.sortedSetAdd(`uid:${uid}:uploads`, Date.now(), relativePath), db.setObjectField(`upload:${md5(relativePath)}`, 'uid', uid)]);
     };
 
     User.deleteUpload = async function (callerUid, uid, uploadNames) {
         if (typeof uploadNames === 'string') {
             uploadNames = [uploadNames];
         } else if (!Array.isArray(uploadNames)) {
-            throw new Error(
-                `[[error:wrong-parameter-type, uploadNames, ${typeof uploadNames}, array]]`,
-            );
+            throw new Error(`[[error:wrong-parameter-type, uploadNames, ${typeof uploadNames}, array]]`);
         }
 
         await _validatePath(uploadNames);
 
-        const [isUsersUpload, isAdminOrGlobalMod] = await Promise.all([
-            db.isSortedSetMembers(`uid:${callerUid}:uploads`, uploadNames),
-            User.isAdminOrGlobalMod(callerUid),
-        ]);
+        const [isUsersUpload, isAdminOrGlobalMod] = await Promise.all([db.isSortedSetMembers(`uid:${callerUid}:uploads`, uploadNames), User.isAdminOrGlobalMod(callerUid)]);
         if (!isAdminOrGlobalMod && !isUsersUpload.every(Boolean)) {
             throw new Error('[[error:no-privileges]]');
         }
@@ -72,42 +54,17 @@ module.exports = function (User) {
 
                 await Promise.all(
                     fullPaths.map(async (fullPath, idx) => {
-                        winston.verbose(
-                            `[user/deleteUpload] Deleting ${uploadNames[idx]}`,
-                        );
-                        await Promise.all([
-                            file.delete(fullPath),
-                            file.delete(
-                                file.appendToFileName(fullPath, '-resized'),
-                            ),
-                        ]);
-                        await Promise.all([
-                            db.sortedSetRemove(
-                                `uid:${uid}:uploads`,
-                                uploadNames[idx],
-                            ),
-                            db.delete(`upload:${md5(uploadNames[idx])}`),
-                        ]);
-                    }),
+                        winston.verbose(`[user/deleteUpload] Deleting ${uploadNames[idx]}`);
+                        await Promise.all([file.delete(fullPath), file.delete(file.appendToFileName(fullPath, '-resized'))]);
+                        await Promise.all([db.sortedSetRemove(`uid:${uid}:uploads`, uploadNames[idx]), db.delete(`upload:${md5(uploadNames[idx])}`)]);
+                    })
                 );
 
                 // Dissociate the upload from pids, if any
-                const pids = await db.getSortedSetsMembers(
-                    uploadNames.map(
-                        relativePath => `upload:${md5(relativePath)}:pids`,
-                    ),
-                );
-                await Promise.all(
-                    pids.map(async (pids, idx) =>
-                        Promise.all(
-                            pids.map(async pid =>
-                                posts.uploads.dissociate(pid, uploadNames[idx]),
-                            ),
-                        ),
-                    ),
-                );
+                const pids = await db.getSortedSetsMembers(uploadNames.map(relativePath => `upload:${md5(relativePath)}:pids`));
+                await Promise.all(pids.map(async (pids, idx) => Promise.all(pids.map(async pid => posts.uploads.dissociate(pid, uploadNames[idx])))));
             },
-            { batch: 50 },
+            { batch: 50 }
         );
     };
 
@@ -123,7 +80,7 @@ module.exports = function (User) {
 
                 setImmediate(next);
             },
-            { batch: 100 },
+            { batch: 100 }
         );
     };
 };
